refactor: deprecate special SYMFONY__ environment vars
The special SYMFONY__ environment variables have been deprecated in
Symfony 3.3 (and will be removed in 4.0) as there is now support for
runtime environment variables in Symfony.

This commit allows for different environment variables to be used as the
first step in removing the old variables.  A later step should remove
the `parameters.yml` file entirely and have the configuration be solely
from passed environment variables (or .env file for development).

ref: [Deprecation blog post](https://symfony.com/blog/new-in-symfony-3-3-deprecated-the-special-symfony-environment-variables)
ref: symfony/symfony#21889