feat: show identifier instead of API URL
The URI for objects is now consitently of the form
"<host>/uri/<identifier>" instead of using the API URL, unless
explicitly something else, such as from another server.

The UI will link to the consistent form even for objects with another
explict URI.

Going to the URI page will display a simple table with the values of
the object.